# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OfferOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns (workspace-scoped, optional batch filter)
      parameters:
        - name: workspace_id
          in: query
          required: true
          schema:
            type: integer
        - name: batch_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a manual campaign row
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Batch not found in workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate campaign for batch+platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /campaigns/{id}:
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign row (status flips emit CampaignStatusChanged)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /daily-reports:
    get:
      operationId: listDailyReports
      tags: [daily-reports]
      summary: List daily reports
      parameters:
        - name: employee_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: workspace_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of daily reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReport"

    post:
      operationId: createDailyReport
      tags: [daily-reports]
      summary: Create daily report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyReportBody"
      responses:
        "201":
          description: Daily report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /daily-reports/{id}:
    get:
      operationId: getDailyReport
      tags: [daily-reports]
      summary: Get daily report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"

    patch:
      operationId: updateDailyReport
      tags: [daily-reports]
      summary: Update daily report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyReportBody"
      responses:
        "200":
          description: Updated daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /weekly-reports:
    get:
      operationId: listWeeklyReports
      tags: [daily-reports]
      summary: List weekly aggregated reports
      parameters:
        - name: employee_id
          in: query
          schema:
            type: integer
        - name: week_start
          in: query
          schema:
            type: string
        - name: workspace_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of weekly reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyReport"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DailyReport:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        reportDate:
          type: string
        offersUploaded:
          type: integer
        batchesCreated:
          type: integer
        batchesTested:
          type: integer
        campaignsMovedToMain:
          type: integer
        campaignsClosed:
          type: integer
        notes:
          type: ["string", "null"]
        workspaceId:
          type: ["integer", "null"]
        createdAt:
          type: string
        employeeName:
          type: ["string", "null"]
      required:
        - id
        - employeeId
        - reportDate
        - offersUploaded
        - batchesCreated
        - batchesTested
        - campaignsMovedToMain
        - campaignsClosed
        - createdAt
    CreateDailyReportBody:
      type: object
      properties:
        employeeId:
          type: integer
        reportDate:
          type: string
        offersUploaded:
          type: integer
        batchesCreated:
          type: integer
        batchesTested:
          type: integer
        campaignsMovedToMain:
          type: integer
        campaignsClosed:
          type: integer
        notes:
          type: ["string", "null"]
        workspaceId:
          type: ["integer", "null"]
      required:
        - employeeId
        - reportDate
        - offersUploaded
        - batchesCreated
        - batchesTested
        - campaignsMovedToMain
        - campaignsClosed
    UpdateDailyReportBody:
      type: object
      properties:
        reportDate:
          type: string
        offersUploaded:
          type: integer
        batchesCreated:
          type: integer
        batchesTested:
          type: integer
        campaignsMovedToMain:
          type: integer
        campaignsClosed:
          type: integer
        notes:
          type: ["string", "null"]
    WeeklyReport:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        weekStart:
          type: string
        weekEnd:
          type: string
        totalOffersUploaded:
          type: integer
        totalBatchesCreated:
          type: integer
        totalBatchesTested:
          type: integer
        totalCampaignsMovedToMain:
          type: integer
        totalCampaignsClosed:
          type: integer
        reportCount:
          type: integer
      required:
        - employeeId
        - employeeName
        - weekStart
        - weekEnd
        - totalOffersUploaded
        - totalBatchesCreated
        - totalBatchesTested
        - totalCampaignsMovedToMain
        - totalCampaignsClosed
        - reportCount
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        batchId:
          type: integer
        platform:
          type: string
          enum: [ios, android]
        campaignName:
          type: string
        trafficSourceId:
          type: ["integer", "null"]
        campaignUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, ready, live, tested, closed]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - workspaceId
        - batchId
        - platform
        - campaignName
        - status
        - createdAt
        - updatedAt
...
    CreateCampaignBody:
      type: object
      properties:
        workspaceId:
          type: integer
        batchId:
          type: integer
        platform:
          type: string
          enum: [ios, android]
        campaignName:
          type: string
        trafficSourceId:
          type: ["integer", "null"]
        campaignUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, ready, live, tested, closed]
      required:
        - workspaceId
        - batchId
        - platform
        - campaignName
    UpdateCampaignBody:
      type: object
      properties:
        campaignName:
          type: string
        trafficSourceId:
          type: ["integer", "null"]
        campaignUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, ready, live, tested, closed]
...